% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-mesh-file.R
\name{read_mesh_file}
\alias{read_mesh_file}
\title{Download and parse an ASCII MeSH File}
\usage{
read_mesh_file(fil, wide = FALSE)
}
\arguments{
\item{fil}{filename of a local (`.bin`) ASCII MeSH file to parse or downlaod
and parse. `fil` will be checked to see if it exists locally and if not
the `nlmpubs.nlm.nih.gov` prefix FTP URL will be added to the file
and the file will be read into memory before parsing.}

\item{wide}{if `TRUE` then the repeating column names will be made unique and
spread out into a wide data frame; otherwise a data frame with nested
columns will be returned.}
}
\description{
If the string passed in matches a local filename it will be read in locally,
otherwise it will be downloaded. NOTE: large files take a while.
}
