% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatamesh-package.R
\docType{package}
\name{whatamesh}
\alias{whatamesh}
\alias{whatamesh-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/whatamesh}
\item BugReports: \url{https://gitlab.com/hrbrmstr/whatamesh/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
